pyproject.toml
[tool.poetry]
name = "fairdm-rest-api"
version = "0.1.0"
description = "A REST API implementation for FAIR-DM"
authors = ["Sam <samuel.scott.jennings@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "fairdm_api", from = "src" }
]


[tool.poetry.dependencies]
python = "^3.11"
dj-rest-auth = "^7.0.1"
django-parler-rest = "^2.2"
django-rest-polymorphic = "^0.1.10"
djangorestframework = "^3.16.1"
djangorestframework-gis = "^1.2.0"
rest-pandas = "^1.1.0"
drf-access-policy = "^1.5.0"
drf-flex-fields = "^1.0.2"
drf-nested-routers = "^0.94.2"
drf-orjson-renderer = "^1.7.3"
drf-schema-adapter = "^3.0.6"
drf-spectacular = {extras = ["sidecar"], version = "^0.28.0"}
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
fairdm-dev-tools = {git = "https://github.com/FAIR-DM/dev-tools"}
shot-scraper = "^1.8"
fairdm = {git = "https://github.com/FAIR-DM/fairdm", rev = "development"}
fairdm-demo = {git = "https://github.com/FAIR-DM/demo-portal"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
